+++ deepmd/pt/utils/nlist.py
@@ -148,7 +148,13 @@ def _trim_mask_distinguish_nlist(
         nlist = nlist[:, :, :nsel]
     else:
         rr = torch.cat(
-            [rr, torch.ones([batch_size, nloc, nsel - nnei], device=rr.device) + rcut],  # pylint: disable=no-explicit-dtype
+            [
+                rr,
+                torch.ones(
+                    [batch_size, nloc, nsel - nnei], device=rr.device, dtype=rr.dtype
+                )
+                + rcut,
+            ],
             dim=-1,
         )
         nlist = torch.cat(
@@ -428,7 +434,10 @@ def extend_coord_with_ghosts(
     """
     device = coord.device
     nf, nloc = atype.shape
-    aidx = torch.tile(torch.arange(nloc, device=device).unsqueeze(0), [nf, 1])  # pylint: disable=no-explicit-dtype
+    # int64 for index
+    aidx = torch.tile(
+        torch.arange(nloc, device=device, dtype=torch.int64).unsqueeze(0), [nf, 1]
+    )
     if cell is None:
         nall = nloc
         extend_coord = coord.clone()
@@ -447,9 +456,15 @@ def extend_coord_with_ghosts(
         # 3
         nbuff = torch.amax(nbuff, dim=0)  # faster than torch.max
         nbuff_cpu = nbuff.cpu()
-        xi = torch.arange(-nbuff_cpu[0], nbuff_cpu[0] + 1, 1, device="cpu")  # pylint: disable=no-explicit-dtype
-        yi = torch.arange(-nbuff_cpu[1], nbuff_cpu[1] + 1, 1, device="cpu")  # pylint: disable=no-explicit-dtype
-        zi = torch.arange(-nbuff_cpu[2], nbuff_cpu[2] + 1, 1, device="cpu")  # pylint: disable=no-explicit-dtype
+        xi = torch.arange(
+            -nbuff_cpu[0], nbuff_cpu[0] + 1, 1, device="cpu", dtype=nbuff_cpu.dtype
+        )
+        yi = torch.arange(
+            -nbuff_cpu[1], nbuff_cpu[1] + 1, 1, device="cpu", dtype=nbuff_cpu.dtype
+        )
+        zi = torch.arange(
+            -nbuff_cpu[2], nbuff_cpu[2] + 1, 1, device="cpu", dtype=nbuff_cpu.dtype
+        )
         eye_3 = torch.eye(3, dtype=env.GLOBAL_PT_FLOAT_PRECISION, device="cpu")
         xyz = xi.view(-1, 1, 1, 1) * eye_3[0]
         xyz = xyz + yi.view(1, -1, 1, 1) * eye_3[1]
